Module epic_wallet_controller::controller

Controller for wallet.. instantiates and handles listeners (or single-run invocations) as needed.

Structs

EPIC_OWNER_BASIC_REALM

ForeignAPIHandlerV2

V2 API Handler/Wrapper for foreign functions

OwnerAPIHandlerV2

V2 API Handler/Wrapper for owner functions

OwnerAPIHandlerV3

V3 API Handler/Wrapper for owner functions, which include a secure mode + lifecycle functions

OwnerV3Helpers

Functions

foreign_listener

Listener version, providing same API but listening for requests on a port and wrapping the calls

foreign_single_use

Instantiate wallet Foreign API for a single-use (command line) call Return a function containing a loaded API context to call

owner_listener

Listener version, providing same API but listening for requests on a port and wrapping the calls Note keychain mask is only provided here in case the foreign listener is also being used in the same wallet instance

owner_single_use

Instantiate wallet Owner API for a single-use (command line) call Return a function containing a loaded API context to call